Fail install.sh and uninstall.sh if no repository provided.

+++ uninstall.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 
+if [ -z "$1" ]; then
+	echo "No repository supplied."
+	exit 1
+fi
+
 directory="$HOME/workspace/$1/.git/hooks"
 
 rm -f $directory/pre-commit
